annotator/grounded-sam/requirements.txt
addict
diffusers
gradio
huggingface_hub
matplotlib
numpy
onnxruntime
opencv-python
Pillow
pycocotools
PyYAML
requests
setuptools
supervision
termcolor
timm
torch
torchvision
transformers
yapf
tqdm
